requirements-server.txt
gradio
uvicorn
fastapi
pydantic_settings
slowapi
requests
python-multipart
pillow
pypdf2
pdf2image